requirements.txt

aiogram>=3.6.0
python-dotenv>=1.0.1
pytest>=8.0.0
pytest-asyncio>=0.23.5
